backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
opencv-python>=4.8.0
face-recognition>=1.3.0
chromadb>=0.4.0
groq>=0.4.0
tiktoken>=0.5.0
pydantic>=2.5.0
python-dotenv>=1.0.0
aiofiles>=23.2.0
httpx>=0.25.0
sounddevice>=0.4.6
numpy>=1.24.0
webrtcvad-wheels>=2.0.10
edge-tts>=6.1.0
openai-whisper>=20231117
sarvamai>=0.1.26a2
